% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{aecdf}
\alias{aecdf}
\title{Approximate empirical commulative distribution function}
\usage{
aecdf(dnull, symmetric = FALSE, n = 100)
}
\arguments{
\item{dnull}{Numerical vector representing the null model}

\item{symmetric}{Logical, whether the distribution should betreated as symmetric around zero and only one tail should be approximated}

\item{n}{Integer indicating the number of points to evaluate the empirical cummulative probability function}
}
\value{
function with two parameters, \code{x} and \code{alternative}
}
\description{
This function generates an empirical null model that computes a normalized statistics and p-value
}
