% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append2tips.R
\name{append2tips}
\alias{append2tips}
\title{Append Symbols or Text to Tip Labels}
\usage{
append2tips(phy, tips, offset, align = FALSE, grid = FALSE,
  col = "red", text = NULL, pch = NULL, cex, legend,
  legend.srt = 90, legend.cex = cex, ...)
}
\arguments{
\item{phy}{An object of class \code{\link[ape]{phylo}}.}

\item{tips}{A vector of mode \code{"numeric"}, giving the tip numbers to
which append text or symbols. If missing, all tips are used.}

\item{offset}{Used to control the distance of text or symbols from
the tips of the phylogenetic tree. Can either be of mode
\code{"numeric"} or \code{"character"}. In the latter case the width of the
string is calculated and used as offset.}

\item{align}{Logical, if \code{TRUE} text or symbols will be aligned.}

\item{grid}{Logial, if \code{TRUE} dotted lines are plotted as a horizontal
grid to aid visual orientation.}

\item{col}{A character string, giving the color for text or symbols.}

\item{text}{A character string, providing text to be appended to the tip
labels.}

\item{pch}{An integer, specifying the symbol used (see \code{\link{points}}).}

\item{cex}{A real number, giving the \strong{c}haracter \strong{ex}pension
(see \code{\link{par}}). If missing, \code{cex} is taking from the previous
call of \code{\link[ape]{plot.phylo}}.}

\item{legend}{A character string to be used as a column legend.}

\item{legend.srt}{A vector of mode \code{"numeric"} to rotate \code{legend}.}

\item{legend.cex}{A real number, giving the \strong{c}haracter
\strong{ex}pension for the legend.}

\item{...}{Further graphical parameters (from par), such as srt, family and
xpd.}
}
\description{
Add text or symbols to the tip labels of a phylogenetic tree.
}
\details{
Currently, \code{append2tips} is will work only for
  \code{\link[ape]{plot.phylo}} using types \code{"phylogram"} or
  \code{"cladogram"} and directions \code{"rightwards"} or
  \code{"leftwards"}.
}
\examples{
# phylogeny of true vipers
# ------------------------
#data(viperidae)
#plot.phylo(viperidae, cex = 0.9, label.offset = 0.01)

# append geographical origin of accessions:
# -----------------------------------------
#txt <- origin[, 2]
#append2tips(viperidae, text = txt)

# If the appended labels - like in the example above -
# exceed the plotting region, a warning is issued and it
# is suggested a new value for x.lim:
# -----------------------------------
#plot.phylo(viperidae, cex = 0.9, label.offset = 0.01, x.lim = 1.1)
#append2tips(viperidae, text = txt)

# a lot of combinations are possible:
# -----------------------------------
#plot.phylo(viperidae, cex = 0.8, label.offset = 0.02, x.lim = 1.5)
#NTips <- length(viperidae$tip.label)
#append2tips(viperidae, text =
#	paste("A", 1:NTips, sep = "-"), col = "orange")
#append2tips(viperidae, tips <- c(5, 7:9, 14),
#	text = rep("xxx", 5), col = "slateblue")
#append2tips(viperidae, pch = 23, col = rainbow(NTips))
#append2tips(viperidae, pch = 21, col = "grey50",
#   align = TRUE, grid = TRUE)

# ... it also works when actual tip labels are omitted:
# -----------------------------------------------------
#plot.phylo(viperidae, show.tip.label = FALSE, x.lim = 40,
#	cex = 0.8, type = "clado", use.edge.length = FALSE)
#txt <- origin[, 2]
#append2tips(viperidae, text = txt)

# ... it can be combined with taxo.fonts:
# -----------------------------------------------------
#plot.phylo(viperidae, show.tip.label = FALSE, x.lim = 1.5)
#taxo.fonts(viperidae)
#txt <- origin[, 2]
#append2tips(viperidae, text = txt, offset = 0.1, grid = TRUE)
}
\seealso{
\code{\link{plot.phylo}}, \code{\link{tiplabels}},
  \code{\link{mixedFontBinom}}
}
