% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigtree.R
\name{bigtree}
\alias{bigtree}
\title{PDF of Large Phylogenies}
\usage{
bigtree(phy, lo = 5, height = "auto", cex = 0.1, edge.label,
  view = TRUE, save = FALSE, file = "bigtree.pdf")
}
\arguments{
\item{phy}{An object of class \code{\link[ape]{phylo}}.}

\item{lo}{Numeric, giving the label offset.}

\item{height}{Either \code{"auto"} or numeric giving the PDF height in
inches.}

\item{cex}{Numeric, giving the character expansion for the tip labels.}

\item{edge.label}{A data.frame with two columns; the taxon names in the first
column (\code{taxon}) are drawn on the edge numbers given by the second
column (\code{edge}).}

\item{view}{Logical, idicating if the PDF should be opened immediately in the
standard PDF viewer.}

\item{save}{Logical, indicating if the PDF file should be written to
\code{file}.}

\item{file}{A character string giving the filename; only has an effect if
\code{save = TRUE}.}
}
\value{
None.
}
\description{
Easily and quickly produce a PDF file showing a large phylogeny.
}
\examples{
tr <- rtree(1000)
bigtree(tr, save = FALSE)
}
\seealso{
\code{\link{plot.phylo}}
}
