% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msviper.r
\name{bootstrapmsviper}
\alias{bootstrapmsviper}
\title{msviper bootstraps integration}
\usage{
bootstrapmsviper(mobj, method = c("mean", "median", "mode"))
}
\arguments{
\item{mobj}{msviper object}

\item{method}{Character string indicating the method to use, either mean, median or mode}
}
\value{
msviper object
}
\description{
This function integrates the bootstrap msviper results
}
\examples{
data(bcellViper, package="bcellViper")
sig <- bootstrapTtest(dset, "description", c("CB", "CC"), "N")
mra <- msviper(sig, regulon)
plot(mra, cex=.7)
}
\seealso{
\code{\link{msviper}}
}
