% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{distMode}
\alias{distMode}
\title{Mode of continuous distributions}
\usage{
distMode(x, adj = 1)
}
\arguments{
\item{x}{Numeric data vector}

\item{adj}{Number indicating the adjustment for the kernel bandwidth}
}
\value{
Number
}
\description{
This function computes the mode for continuous distributions
}
\examples{
data(bcellViper, package="bcellViper")
d1 <- exprs(dset)
mean(d1[, 1])
median(d1[, 1])
distMode(d1[, 1])
plot(density(d1[, 1]))
abline(v=c(mean(d1[, 1]), median(d1[, 1]), distMode(d1[, 1])), col=c("green", "red", "blue"))
legend("topleft", c("Mean", "Median", "Mode"), col=c("green", "red", "blue"), lwd=4)
}
