% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{fcvarna}
\alias{fcvarna}
\title{Variance of columns for arrays with NA values}
\usage{
fcvarna(x)
}
\arguments{
\item{x}{Numeric matrix}
}
\value{
1-column matrix with the variance by column results
}
\description{
This function computes the variance by columns ignoring NA values
}
\examples{
data(bcellViper, package="bcellViper")
tmp <- exprs(dset)[, 1:10]
tmp[round(runif(100, 1, length(tmp)))] <- NA
fcvarna(tmp)
}
