% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillClades.R
\name{fillClades}
\alias{fillClades}
\title{Highlight Clades in a Phylogeny}
\usage{
fillClades(phy, nodes, col, text, text.cex = 1, open.angle = 0)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{nodes}{A vector of mode \code{"numeric"} giving the most recent common
ancestor (MRCA) node of the clades to be highlighted. You can use
\code{\link{noi}} to obtain this vector.}

\item{col}{A vector of mode \code{"character"} giving the color(s) of the
clades; will be receycled if appropriate; default is \code{"grey"}.}

\item{text}{A vector of mode \code{"character"}, giving (optional) text
labels for clades; must be of the same length as \code{nodes}. The labels
will be positioned right justified and vertically centered.}

\item{text.cex}{Numeric, giving the character expansion (size) of the text
labels.}

\item{open.angle}{Numeric, giving the the angle left blank in degrees and has
to match the value used in \code{\link{plot.phylo}} when choosing a
circular plot type (\code{"fan"} or \code{"radial"}); for other plot types
this has no effect.}
}
\description{
Highlights one or more monophyletic groups (clades) in a
  phylogeny by filling them with color.
}
\details{
On 2018-01-17, I added the optional plotting of labels. To this end,
  in non-ultrametric trees the tip-ward side of the polygons has to be
  aligned. This has only been tested for rightward phylograms. Maybe boxes
  should be extended vertically a bit to include tiplabels.And maybe the code
  should be organized a bit more systematically.
}
