% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{filterColMatrix}
\alias{filterColMatrix}
\title{Filter for columns of a matrix with no loss of col and row names}
\usage{
filterColMatrix(x, filter)
}
\arguments{
\item{x}{Matrix}

\item{filter}{Logical or numerical index of columns}
}
\value{
Matrix
}
\description{
This function filters the columns of a matrix returning always a two dimensional matrix
}
