% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{loadExpset}
\alias{loadExpset}
\title{Loading expression sets}
\usage{
loadExpset(filename)
}
\arguments{
\item{filename}{Character string indicating the name of the expression file}
}
\value{
List containing a numeric matrix of expression data with samples in columns and probes in rows; and a vector of gene mapping annotations
}
\description{
This function load an expression file into a matrix
}
