% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.r
\name{msviperClass}
\alias{msviperClass}
\title{msVIPER class}
\usage{
msviperClass(nes, signature, regulon, nullmodel = NULL)
}
\arguments{
\item{nes}{Numeric vector of NES values}

\item{signature}{Numeric vector of gene expression signature}

\item{regulon}{Instance of class regulon}

\item{nullmodel}{Optional matrix containing the signatures for the null model}
}
\value{
msviper class object
}
\description{
This function generates an instance of the msviper class from a signature, NES signature and regulon object
}
\examples{
data(bcellViper, package="bcellViper")
sig <- rowTtest(dset, "description", c("CB", "CC"), "N")$statistic
mra <- msviper(sig, regulon)
mra1 <- msviperClass(mra$es$nes, sig, regulon)
summary(mra1)
plot(mra1)
}
