% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msviper.r
\name{msviperSynergy}
\alias{msviperSynergy}
\title{msviper synergy analysis}
\usage{
msviperSynergy(mobj, per = 1000, seed = 1, verbose = TRUE)
}
\arguments{
\item{mobj}{msviper object containing combinatorial regulation results generated by \code{msviperCombinatorial}}

\item{per}{Integer indicating the number of permutations}

\item{seed}{Integer indicating the seed for the permutations, 0 for disable it}

\item{verbose}{Logical, whether progression messages should be printed in the terminal}
}
\value{
Updated msviper object containing the sygergy p-value
}
\description{
This function performs a synergy analysis for combinatorial regulation
}
\examples{
data(bcellViper, package="bcellViper")
sig <- rowTtest(dset, "description", c("CB", "CC"), "N")$statistic
dnull <- ttestNull(dset, "description", c("CB", "CC"), "N")
mra <- msviper(sig, regulon, dnull)
mra <- msviperCombinatorial(mra, 50)
mra <- msviperSynergy(mra)
summary(mra)
}
\seealso{
\code{\link{msviper}}
}

