% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPhyloUpon.R
\name{plotPhyloUpon}
\alias{plotPhyloUpon}
\title{Plot Phylogenies}
\usage{
plotPhyloUpon(x, type = "phylogram", use.edge.length = TRUE,
  node.pos = NULL, show.tip.label = TRUE, show.node.label = FALSE,
  edge.color = "black", edge.width = 1, edge.lty = 1, font = 3,
  cex = par("cex"), adj = NULL, srt = 0, no.margin = FALSE,
  root.edge = FALSE, label.offset = 0, underscore = FALSE,
  x.lim = NULL, y.lim = NULL, direction = "rightwards",
  lab4ut = NULL, tip.color = "black", plot = TRUE, rotate.tree = 0,
  open.angle = 0, node.depth = 1, align.tip.label = FALSE, ...)
}
\description{
Plot phylogenetic trees exactly like \code{\link{plot.phylo}}, 
  with one important difference: The plot will be added to existing plots in
  the graphical device.
}
