% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/internal.r
\name{pwea3NULLf}
\alias{pwea3NULLf}
\title{Null model function}
\usage{
pwea3NULLf(pwnull, verbose = TRUE)
}
\arguments{
\item{pwnull}{Object generated by \code{pwea3NULLgroups} function}

\item{verbose}{Logical, whether progression messages should be printed in the terminal}
}
\value{
List of function to compute NES and p-value
}
\description{
This function generates the NULL model function, which computes the normalized enrichment score and associated p-value
}

