% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureDistance.r
\name{scaleGroups}
\alias{scaleGroups}
\title{Signatures with grouping variable}
\usage{
scaleGroups(x, groups)
}
\arguments{
\item{x}{Numerical matrix with genes in rows and samples in columns}

\item{groups}{Vector of same length as columns has the dset containing the labels for grouping the samples}
}
\value{
Numeric matrix of signatures (z-scores) with genes in rows and groups in columns
}
\description{
\code{scaleGroups} compares each group vs. the remaining groups using a Student's t-test
}
\details{
This function compute signatures using groups information
}
\examples{
data(bcellViper, package="bcellViper")
res <- scaleGroups(exprs(dset)[, 1:20], rep(1:4, rep(5, 4)))
res[1:5, ]
}
