% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases.r
\name{signatureDistance-class}
\alias{signatureDistance-class}
\title{signatureDistance}
\description{
This class contains the results generated by \code{signatureDistance} function.
}
\section{Slots}{

Matrix of class \code{numeric} containing the similarity scores
}

