% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.r
\name{sigT}
\alias{sigT}
\title{Sigmoid transformation}
\usage{
sigT(x, slope = 20, inflection = 0.5)
}
\arguments{
\item{x}{Numeric vector}

\item{slope}{Number indicating the slope at the inflection point}

\item{inflection}{Number indicating the inflection point}
}
\value{
Numeric vector
}
\description{
This function transforms a numeric vector using a sigmoid function
}
