% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msviper.r
\name{summary.msviper}
\alias{summary.msviper}
\title{List msviper results}
\usage{
\method{summary}{msviper}(object, mrs = 10, ...)
}
\arguments{
\item{object}{msviper object}

\item{mrs}{Either number of top MRs to report or vector containing the genes to display}

\item{...}{Given for compatibility with the summary generic function}
}
\value{
Data.frame with results
}
\description{
This function generates a table of msviper results
}
