\name{viper-package}
\alias{viper-package}
\alias{viper}
\docType{package}
\title{
Visualization of Phylogenies and Evolution in R
}
\description{
The package provides functions for plotting phylogenetic data.
}
\details{
\tabular{ll}{
Package: \tab viper\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
 Date: \tab 2014-01-15\cr
License: \tab GPL (>= 2)\cr
}
\code{viper} provides functions that are designed to work with phylogenetic trees represented as objects of class \code{\link[ape]{phylo}} in R. 
}
\author{
Christoph Heibl
Maintainer: Christoph Heibl <christoph.heibl@gmx.net>
}
%\references{
%~~ Literature or other references for background information ~~
%}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
