% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases.r
\docType{class}
\name{viperSignature-class}
\alias{viperSignature-class}
\title{viperSignature}
\description{
This class contains the results produced by the \code{viperSignature} function
}
\section{Slots}{

 \describe{
   \item{\code{signature}:}{Matrix of class \code{numeric} with genes in rows and samples in columns containing the gene expression signatures}
   \item{\code{nullmodel}:}{Matrix of class \code{numeric} with genes in rows and permutations in columns ontaining the sample-permutation based signatures to be used as NULL model}
 }
}

