\name{viper}
\alias{viper}
\title{VIPER}
\usage{
  viper(eset, regulon, dnull = NULL, tw = 0.5, lw = 1,
    nes = TRUE,
    method = c("scale", "rank", "mad", "ttest", "none"),
    minsize = 25, adaptive.size = FALSE,
    eset.filter = TRUE, nes.method = "analytic",
    verbose = TRUE)
}
\arguments{
  \item{eset}{ExpressionSet object or Numeric matrix
  containing the expression data or gene expression
  signatures, with samples in columns and genes in rows}

  \item{regulon}{Object of class regulon}

  \item{dnull}{Numeric matrix for the null model, usually
  generated by \code{nullTtest}}

  \item{tw}{Number indicating the power transformation for
  target weight}

  \item{lw}{Number indicating the power transformation for
  the likelihood weight}

  \item{nes}{Logical, whether the enrichment score reported
  should be normalized}

  \item{method}{Character string indicating the method for
  computing the single samples signature, either scale,
  rank, mad, ttest or none}

  \item{minsize}{Integer indicating the minimum number of
  targets allowed per regulon}

  \item{adaptive.size}{Logical, whether the weighting
  scores should be taken into account for computing the
  regulon size}

  \item{eset.filter}{Logical, whether the dataset should be
  limited only to the genes represented in the interactome}

  \item{nes.method}{Character string indicating the method
  for estimating the NES, either analytical, empirical or
  fit. Fit method is recommended only for large datasets
  (>500 samples)}

  \item{verbose}{Logical, whether progression messages
  should be printed in the terminal}
}
\value{
  A matrix of inferred activity for each regulator gene in
  the network across all samples
}
\description{
  This function performs Virtual Inference of
  Protein-activity by Enriched Regulon analysis
}
\examples{
data(bcellViper, package="bcellViper")
d1 <- exprs(dset)
res <- viper(d1, regulon)
dim(d1)
d1[1:5, 1:5]
regulon
dim(res)
res[1:5, 1:5]
}
\seealso{
  \code{\link{msviper}}
}

